contribution/scripts/requirements.txt
# Python Dependencies for EIP-8004 Contribution Scripts
# Week 2: Data Collection & Analysis

# Blockchain interaction
web3>=6.0.0
eth-account>=0.8.0

# AWS integration (for secrets)
boto3>=1.28.0

# Configuration
python-dotenv>=1.0.0

# Data analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter for interactive analysis
jupyter>=1.0.0
ipykernel>=6.25.0

# HTTP requests
requests>=2.31.0

# Progress bars for scripts
tqdm>=4.66.0

# JSON manipulation
jsonschema>=4.19.0

# Date/time utilities
python-dateutil>=2.8.2

# Network graph visualization
networkx>=3.1
# Optional: For interactive graphs
# plotly>=5.17.0
